refactor(github): extract pure Pr*/GitHub* mappers into githubMappers.ts
Moves the 11 payload-shaping mappers (mapPrState, aggregateChecksStatus,
mapPrData, mapPrAuthor, mapPrCommit, mapPrComment, mapPrReview, mapStatusCheck,
mapPrDetails, parseGhAuthAccounts, mapGitHubApiRepo) out of github.ts into a new
githubMappers.ts leaf module. github.ts imports the four it calls directly
(mapPrData/mapPrDetails/parseGhAuthAccounts/mapGitHubApiRepo) and re-exports the
three public ones, so the module's public surface is unchanged. github.ts drops
from 1035 to 776 lines and now composes over the mappers instead of co-owning
~250 lines of raw payload handling.

Pure move — no logic edits. github.test.ts (57 cases) green, lint 0/0.
